Move a bunch of common utils from the dashboard folder
