+++ src/app/_index.js
@@ -25,7 +25,7 @@ import { addNewCustomTemplate } from '../common/utils/custom_templates'
 import { dispatchingToStore, makeFlagConsistent } from './makeFlagConsistent'
 import { TEMP_FILES_PATH } from '../file-system/config_paths'
 import { createErrorReport } from '../common/utils/full_error_report'
-import TemplateFetcher from '../dashboard/utils/template_fetcher'
+import TemplateFetcher from '../common/utils/template_fetcher'
 import { machineIdSync } from 'node-machine-id'
 import Listener from './components/listener'
 import Renamer from './components/Renamer'
