+++ src/common/utils/templates_from_firestore.js
