+++ src/common/utils/window_manager.js
